VebFinanc/requirements.txt
# pip install -r requirements.txt
Flask
Flask-BS4==4.5.3.0
Flask-WTF
WTForms
python-dateutil
pyodbc
click
openpyxl